moved into component folders and fixed imports
