+++ frontend/src/components/addReview/AddReviewModal.jsx
@@ -1,8 +1,8 @@
 import React, { useContext, useRef, useState } from 'react';
 import PropTypes from 'prop-types';
 import axios from 'axios';
-import { ModalContext } from '../context/ModalContext';
-import { AuthContext } from '../context/AuthContext';
+import { ModalContext } from '../../context/ModalContext';
+import { AuthContext } from '../../context/AuthContext';
 
 function AddReviewModal({ setIsModal, refresh, setRefresh, toiletId }) {
   const contentRef = useRef();
@@ -34,16 +34,12 @@ function AddReviewModal({ setIsModal, refresh, setRefresh, toiletId }) {
       return;
     }
     try {
-      const response = await axios.post(
-        `/toilets/${toiletId}/review`,
-        body,
-        {
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
+      const response = await axios.post(`/toilets/${toiletId}/review`, body, {
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: `Bearer ${token}`,
+        },
+      });
       setRefresh(!refresh);
       pushModal({
         message: response.data.message,
