+++ frontend/src/components/review/Review.jsx
@@ -1,11 +1,9 @@
 /* eslint-disable react/no-unused-prop-types */
 import React from 'react';
 import PropTypes from 'prop-types';
-import timeCalculator from '../utils/timeCalculator';
+import timeCalculator from '../../utils/timeCalculator';
 
-function Review({
-  clean, content, author, createdAt,
-}) {
+function Review({ clean, content, author, createdAt }) {
   return (
     <div
       className="flex items-center justify-between gap-6 rounded-md border p-4 px-8"
